// client/src/types/ChatMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChatMessageType } from "./ChatMessageType";

export type ChatMessage = { id: string, campaign_id: string, sender_user_id: string, sender_display_name: string, character_id: string | null, character_name: string | null, message_type: ChatMessageType, content: string, whisper_target_ids: Array<string>, created_at: string, };